Low Pass Filter and Volume Controls
Both Filters Implemented and Most Keyboard Controls as well.  SRC is
still broken and so are the filters.  I think the filters have to do
with the difference equation and can be solved by using pointers for
the last 4 samples to be used in the next round of chunk processing.